LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/inline-event-handler-ordering-expected.txt

CONSOLE MESSAGE: line 18: SyntaxError: Unexpected token '}'
CONSOLE MESSAGE: line 51: SyntaxError: Unexpected token '}'

Harness Error (FAIL), message = SyntaxError: Unexpected token '}'

PASS Inline event handlers retain their ordering when invalid and force-compiled 
PASS Inline event handlers retain their ordering when invalid and force-compiled via dispatch 
PASS Inline event handlers retain their ordering when invalid and lazy-compiled 
